Remove reads with duplicated sequences.
When digital tag is added to the beginning of the sequence, reads with
the same digital tag + sequence of certain length are counted as
amplification duplicates.

The script also checks the quality score of certain length. Since the
output only include one unique read, the quality score information of
other duplicated reads are lost. Filtering for quality score at
downstream analysis will be invalid.